+++ website/static/styles.css
@@ -1,6 +1,6 @@
-body {
+/* body {
     background-color: #e5e5e5;
-}
+} */
 
 h1 {
     font-size: 1.7rem; /* Уменьшите размер шрифта по своему усмотрению */
