+++ website/templates/call_log/calls_log.html
@@ -4,26 +4,12 @@
 
 {% block content %}
 
-<style>
-    /* Стилизация таблицы */
-    table {
-        width: 100%;
-        border-collapse: collapse;
-    }
-    th, td {
-        padding: 8px;
-        text-align: left;
-        border-bottom: 1px solid #ddd;
-    }
-    tr:hover {background-color: #f5f5f5;}
-</style>
-
 <h1>Журнал звонков</h1>
 
 <!-- Фильтр по дате -->
     <div class="row mb-3">
         <div class="col-md-6">
-            <form method="GET" action="{{ url_for('calls_log') }}">
+            <form method="GET" action="{{ url_for('call_log_bp.show_log') }}">
                 <div class="row gx-2 align-items-center">
                     <div class="col-auto">
                         <label for="fromdt" class="col-form-label">От:</label>
@@ -72,7 +58,7 @@
             <td>{{ call.call_status }}</td>
             <td>
                 {% if call.record_file %}
-                <a href="{{ url_for('get_record_file', id=call.id) }}" target="_blank">Запись</a>
+                <a href="{{ url_for('call_log_bp.get_record_file', id=call.id) }}" target="_blank">Запись</a>
                 {% endif %}
             </td>
             <td>
@@ -93,17 +79,17 @@
     <ul class="pagination">
         <!-- Первая страница -->
         <li class="page-item {% if not pagination.has_prev %}disabled{% endif %}">
-            <a class="page-link" href="{{ url_for('calls_log', fromdt=fromdt, todt=todt, page=pagination.prev_num) }}">«</a>
+            <a class="page-link" href="{{ url_for('call_log_bp.show_log', fromdt=fromdt, todt=todt, page=pagination.prev_num) }}">«</a>
         </li>
         <!-- Нумерация страниц -->
         {% for page_num in pagination.iter_pages() %}
         <li class="page-item {% if page_num == pagination.page %}active{% endif %}">
-            <a class="page-link" href="{{ url_for('calls_log', fromdt=fromdt, todt=todt, page=page_num) }}">{{ page_num }}</a>
+            <a class="page-link" href="{{ url_for('call_log_bp.show_log', fromdt=fromdt, todt=todt, page=page_num) }}">{{ page_num }}</a>
         </li>
         {% endfor %}
         <!-- Последняя страница -->
         <li class="page-item {% if not pagination.has_next %}disabled{% endif %}">
-            <a class="page-link" href="{{ url_for('calls_log', fromdt=fromdt, todt=todt, page=pagination.next_num) }}">»</a>
+            <a class="page-link" href="{{ url_for('call_log_bp.show_log', fromdt=fromdt, todt=todt, page=pagination.next_num) }}">»</a>
         </li>
 </ul>
 </nav>
