+++ website/templates/users/users.html
@@ -5,7 +5,6 @@
 
 {% block content %}
 <h1>Список абонентов</h1>
-
 <div class="d-flex justify-content-between align-items-center mb-3">
     <form method="GET">
         <div class="input-group">
